More training features (#61)
* Rough automatic mixed precision implementation (not tested yet)

* Fix bugs in amp implementation (now runs on GPU; have to check if no-amp for tensorreps_transform is necessary)

* Consistently implement ExponentialMovingAverage (ema); was there earlier but not consistently implemented

* Add key float32_matmul_precision (one in highest, high, medium)

* Dont avoid autocast in frame-to-frame conversion

+++ lloca/reps/tensorreps_transform.py
@@ -116,7 +116,7 @@ class TensorRepsTransform(torch.nn.Module):
 
             einsum_string = get_einsum_string(rep.order)
             x_transformed = torch.einsum(
-                einsum_string, *([lframes.matrices] * rep.order), x
+                einsum_string, *([lframes.matrices.to(x.dtype)] * rep.order), x
             )
             output[:, idx_start:idx_end] = x_transformed.reshape(-1, mul_rep.dim)
 
@@ -159,7 +159,9 @@ class TensorRepsTransform(torch.nn.Module):
 
             if order > 0:
                 # apply transformation, then flatten because transformation is done
-                output = torch.einsum("ijk,ilk...->ilj...", lframes.matrices, output)
+                output = torch.einsum(
+                    "ijk,ilk...->ilj...", lframes.matrices.to(output.dtype), output
+                )
                 output = output.flatten(start_dim=1, end_dim=2)
 
         return output
@@ -187,7 +189,7 @@ class TensorRepsTransform(torch.nn.Module):
         vectors = vectors.reshape(tensor.shape[0], -1, 4)
         vectors_transformed = torch.einsum(
             "ijk,ilk->ilj",
-            lframes.matrices,
+            lframes.matrices.to(vectors.dtype),
             vectors,
         )
         output[:, vector_idx_start:vector_idx_end] = vectors_transformed.reshape(
